+++ vault-plugins/flant_iam_auth/io/utils/tests/storage.go
@@ -7,7 +7,7 @@ import (
 	"github.com/hashicorp/vault/sdk/logical"
 	"github.com/stretchr/testify/require"
 
-	"github.com/flant/negentropy/vault-plugins/flant_iam_auth/model"
+	"github.com/flant/negentropy/vault-plugins/flant_iam_auth/repo"
 	"github.com/flant/negentropy/vault-plugins/shared/io"
 	sharedkafka "github.com/flant/negentropy/vault-plugins/shared/kafka"
 )
@@ -18,7 +18,7 @@ func CreateTestStorage(t *testing.T) *io.MemoryStore {
 	mb, err := sharedkafka.NewMessageBroker(context.TODO(), storageView)
 	require.NoError(t, err)
 
-	schema, err := model.GetSchema()
+	schema, err := repo.GetSchema()
 	require.NoError(t, err)
 
 	storage, err := io.NewMemoryStore(schema, mb)
