+++ vault-plugins/flant_iam_auth/path_issue.go
@@ -8,8 +8,7 @@ import (
 	"github.com/hashicorp/vault/sdk/logical"
 
 	iam_repo "github.com/flant/negentropy/vault-plugins/flant_iam/repo"
-	"github.com/flant/negentropy/vault-plugins/flant_iam_auth/model"
-	"github.com/flant/negentropy/vault-plugins/flant_iam_auth/model/repo"
+	repo2 "github.com/flant/negentropy/vault-plugins/flant_iam_auth/repo"
 	"github.com/flant/negentropy/vault-plugins/flant_iam_auth/usecase"
 	backendutils "github.com/flant/negentropy/vault-plugins/shared/backent-utils"
 	jwt "github.com/flant/negentropy/vault-plugins/shared/jwt/usecase"
@@ -101,7 +100,7 @@ func (b *flantIamAuthBackend) pathIssueJwt(ctx context.Context, req *logical.Req
 		return nil, fmt.Errorf("cannot cast 'options' to map[string]interface{}")
 	}
 
-	repo := repo.NewJWTIssueTypeRepo(tnx)
+	repo := repo2.NewJWTIssueTypeRepo(tnx)
 	jwtType, err := repo.Get(name)
 	if err != nil {
 		return nil, err
@@ -169,7 +168,7 @@ func (b *flantIamAuthBackend) pathIssueMultipassJwt(ctx context.Context, req *lo
 	multipassService := &usecase.Multipass{
 		JwtController:    b.jwtController,
 		MultipassRepo:    iam_repo.NewMultipassRepository(tnx),
-		GenMultipassRepo: model.NewMultipassGenerationNumberRepository(tnx),
+		GenMultipassRepo: repo2.NewMultipassGenerationNumberRepository(tnx),
 		Logger:           b.NamedLogger("MultipassNewGen"),
 	}
