+++ vault-plugins/flant_iam_auth/repo/multipassgenerationnumber_test.go
@@ -1,4 +1,4 @@
-package model
+package repo
 
 import (
 	"testing"
