+++ vault-plugins/flant_iam_auth/repo/schema.go
@@ -1,4 +1,4 @@
-package model
+package repo
 
 import (
 	"fmt"
